downloading asset for pre and post upgrade
