+++ cosmovisor/upgrade_test.go
@@ -36,7 +36,7 @@ func (s *upgradeTestSuite) TestCurrentBin() {
 
 	// ensure we cannot set this to an invalid value
 	for _, name := range []string{"missing", "nobin", "noexec"} {
-		s.Require().Error(cfg.SetCurrentUpgrade(cosmovisor.UpgradeInfo{Name: name}), name)
+		s.Require().Error(cfg.SetCurrentUpgrade(cosmovisor.Plan{Name: name}), name)
 
 		currentBin, err := cfg.CurrentBin()
 		s.Require().NoError(err)
@@ -47,7 +47,7 @@ func (s *upgradeTestSuite) TestCurrentBin() {
 	// try a few times to make sure this can be reproduced
 	for _, name := range []string{"chain2", "chain3", "chain2"} {
 		// now set it to a valid upgrade and make sure CurrentBin is now set properly
-		err = cfg.SetCurrentUpgrade(cosmovisor.UpgradeInfo{Name: name})
+		err = cfg.SetCurrentUpgrade(cosmovisor.Plan{Name: name})
 		s.Require().NoError(err)
 		// we should see current point to the new upgrade dir
 		currentBin, err := cfg.CurrentBin()
@@ -66,7 +66,7 @@ func (s *upgradeTestSuite) TestCurrentAlwaysSymlinkToDirectory() {
 	s.Require().Equal(cfg.GenesisBin(), currentBin)
 	s.assertCurrentLink(cfg, "genesis")
 
-	err = cfg.SetCurrentUpgrade(cosmovisor.UpgradeInfo{Name: "chain2"})
+	err = cfg.SetCurrentUpgrade(cosmovisor.Plan{Name: "chain2"})
 	s.Require().NoError(err)
 	currentBin, err = cfg.CurrentBin()
 	s.Require().NoError(err)
@@ -99,7 +99,7 @@ func (s *upgradeTestSuite) TestDoUpgradeNoDownloadUrl() {
 
 	// do upgrade ignores bad files
 	for _, name := range []string{"missing", "nobin", "noexec"} {
-		info := cosmovisor.UpgradeInfo{Name: name}
+		info := cosmovisor.Plan{Name: name}
 		err = cosmovisor.DoUpgrade(cfg, info)
 		s.Require().Error(err, name)
 		currentBin, err := cfg.CurrentBin()
@@ -110,7 +110,7 @@ func (s *upgradeTestSuite) TestDoUpgradeNoDownloadUrl() {
 	// make sure it updates a few times
 	for _, upgrade := range []string{"chain2", "chain3"} {
 		// now set it to a valid upgrade and make sure CurrentBin is now set properly
-		info := cosmovisor.UpgradeInfo{Name: upgrade}
+		info := cosmovisor.Plan{Name: upgrade}
 		err = cosmovisor.DoUpgrade(cfg, info)
 		s.Require().NoError(err)
 		// we should see current point to the new upgrade dir
@@ -125,65 +125,6 @@ func (s *upgradeTestSuite) TestDoUpgradeNoDownloadUrl() {
 func (s *upgradeTestSuite) TestOsArch() {
 	// all download tests will fail if we are not on linux...
 	s.Require().Equal("linux/amd64", cosmovisor.OSArch())
-}
-
-func (s *upgradeTestSuite) TestGetDownloadURL() {
-	// all download tests will fail if we are not on linux...
-	ref, err := filepath.Abs(filepath.FromSlash("./testdata/repo/ref_to_chain3-zip_dir.json"))
-	s.Require().NoError(err)
-	badref, err := filepath.Abs(filepath.FromSlash("./testdata/repo/chain2-zip_bin/autod.zip")) // "./testdata/repo/zip_binary/autod.zip"))
-	s.Require().NoError(err)
-
-	cases := map[string]struct {
-		info string
-		url  string
-		err  string
-	}{
-		"missing": {
-			err: "downloading reference link : invalid source string:",
-		},
-		"follow reference": {
-			info: ref,
-			url:  "https://github.com/cosmos/cosmos-sdk/raw/master/cosmovisor/testdata/repo/chain3-zip_dir/autod.zip?checksum=sha256:8951f52a0aea8617de0ae459a20daf704c29d259c425e60d520e363df0f166b4",
-		},
-		"malformated reference target": {
-			info: badref,
-			err:  "upgrade info doesn't contain binary map",
-		},
-		"missing link": {
-			info: "https://no.such.domain/exists.txt",
-			err:  "dial tcp: lookup no.such.domain: no such host",
-		},
-		"proper binary": {
-			info: `{"binaries": {"linux/amd64": "https://foo.bar/", "windows/amd64": "https://something.else"}}`,
-			url:  "https://foo.bar/",
-		},
-		"any architecture not used": {
-			info: `{"binaries": {"linux/amd64": "https://foo.bar/", "*": "https://something.else"}}`,
-			url:  "https://foo.bar/",
-		},
-		"any architecture used": {
-			info: `{"binaries": {"linux/arm": "https://foo.bar/arm-only", "any": "https://foo.bar/portable"}}`,
-			url:  "https://foo.bar/portable",
-		},
-		"missing binary": {
-			info: `{"binaries": {"linux/arm": "https://foo.bar/"}}`,
-			err:  "cannot find binary for",
-		},
-	}
-
-	for name, tc := range cases {
-		s.Run(name, func() {
-			url, err := cosmovisor.GetDownloadURL(cosmovisor.UpgradeInfo{Info: tc.info})
-			if tc.err != "" {
-				s.Require().Error(err)
-				s.Require().Contains(err.Error(), tc.err)
-			} else {
-				s.Require().NoError(err)
-				s.Require().Equal(tc.url, url)
-			}
-		})
-	}
 }
 
 func (s *upgradeTestSuite) TestDownloadBinary() {
@@ -247,7 +188,7 @@ func (s *upgradeTestSuite) TestDownloadBinary() {
 		}
 
 		upgrade := "amazonas"
-		info := cosmovisor.UpgradeInfo{
+		info := cosmovisor.Plan{
 			Name: upgrade,
 			Info: fmt.Sprintf(`{"binaries":{"%s": "%s"}}`, cosmovisor.OSArch(), url),
 		}
